Implement "best neighbour" restoration
The logic will now try to find a neighbour with the same width or
height, since those would pop at just the right spot after a split or
vsplit. Anything with a different size will be mismatched.

In the particular case when the window is full-width or full-height, we
can try with `botright` and `topleft`, though it's not ideal, since
it'll shunt them to the edge of the screen. It will not work for full
windows in the middle. Still, it's better than not having it.